Demo_Dungeon_Full
Banners on ceiling are where enemy interactions should be (same as initial 2 demos grapple and invisibility respectively).

Let me know asap if the two wood grapple platforms (both sets) are hard to see or understand depending on camera angle. or anything for that matter.

Key is our placeholder pickup for the time being